+++ uCOS-II_Wrapper/uCOS-II/os_cfg.h
@@ -142,7 +142,7 @@
 #define OS_Q_POST_EN              1u   /*     Include code for OSQPost()                               */
 #define OS_Q_POST_FRONT_EN        1u   /*     Include code for OSQPostFront()                          */
 #define OS_Q_POST_OPT_EN          0u   /*     Include code for OSQPostOpt()                            */
-#define OS_Q_QUERY_EN             0u   /*     Include code for OSQQuery()                              */
+#define OS_Q_QUERY_EN             1u   /*     Include code for OSQQuery()                              */
 
 
                                        /* ------------------------ SEMAPHORES ------------------------ */